internal/simd/add.go
// {{{ Copyright (c) Paul R. Tagliamonte <paul@k3xec.com>, 2020
//
// Permission is hereby granted, free of charge, to any person obtaining a copy
// of this software and associated documentation files (the "Software"), to deal
// in the Software without restriction, including without limitation the rights
// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
// copies of the Software, and to permit persons to whom the Software is
// furnished to do so, subject to the following conditions:
//
// The above copyright notice and this permission notice shall be included in
// all copies or substantial portions of the Software.
//
// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
// THE SOFTWARE. }}}

package simd

import (
	"fmt"
)

// AddComplex will add the passed slice of complex numbers in buffer a
// to the values in buffer b, and write the results to buffer c.
//
// The buffers *must* all be the same length, or an error will be returned.
//
// Buffer c may be set to one of a or b safely to preform an in-place addition.
func AddComplex(a, b, c []complex64) error {
	if len(a) != len(b) || len(a) != len(c) {
		return fmt.Errorf("simd.AddComplex: a, b, and c are not the same length")
	}
	addComplex(a, b, c)
	return nil
}

// vim: foldmethod=marker
